Update apply_make_intensity_profile.py
Added the cell_label variable as a parameter in the plot_intensity_profiles_and_cells function and in the get_intensity_profile_function. This allows two functions. If the cell_label == 0 then the user gets to choose a mask as input. Otherwise the function directly plots the selected cell_label.